requirements.txt
discord.py==2.4.0
nltk==3.9.1
pyenchant==3.2.2
